[update]権限チェック

+++ app/controllers/items_controller.rb
@@ -1,7 +1,7 @@
 class ItemsController < ApplicationController
 # load_and_authorize_resource
 def index
-    @items = Item.all.page(params[:page]).per(10)
+    @items = Item.all.page(params[:page]).per(12)
 end
 
 
@@ -45,9 +45,14 @@ def category_items
 end
 
 def destroy
-  item = Item.find(params[:id])
-  item.destroy
-  redirect_to items_path
+  if @item.status == '貸出可'
+    item = Item.find(params[:id])
+    item.destroy
+    redirect_to items_path
+  else
+    flash[:notice] = "その備品は貸出中のため削除できません。"
+    redirect_back(fallback_location: root_path)
+  end
 end
 
 private
